skills/yt-transcribe/requirements.txt
yt-dlp
mlx-whisper
python-pptx
openai-whisper
